Remove redundant trim_wiki file
